requirements.txt
click==8.0.1
Flask==2.0.1
flask-cors
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==10.0.0
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.2
Werkzeug==2.0.1